System in place to now break up conditions
Therapist can now pick the condition they want game to go by: Timer, Ball Hit Goal, or both.

The way it works is the other condition that is off has a very high goal hard coded into the engine so player unlikely to hit it. It isnt ideal but it'll work for this class.